feat: move service_util to separated package
