+++ test/e2e/framework/BUILD
@@ -24,7 +24,6 @@ go_library(
         "pv_util.go",
         "rc_util.go",
         "resource_usage_gatherer.go",
-        "service_util.go",
         "size.go",
         "test_context.go",
         "util.go",
@@ -45,7 +44,6 @@ go_library(
         "//pkg/kubelet/metrics:go_default_library",
         "//pkg/kubelet/sysctl:go_default_library",
         "//pkg/master/ports:go_default_library",
-        "//pkg/registry/core/service/portallocator:go_default_library",
         "//pkg/scheduler/algorithm/predicates:go_default_library",
         "//pkg/scheduler/nodeinfo:go_default_library",
         "//pkg/security/podsecuritypolicy/seccomp:go_default_library",
@@ -92,7 +90,6 @@ go_library(
         "//staging/src/k8s.io/client-go/tools/clientcmd/api:go_default_library",
         "//staging/src/k8s.io/client-go/tools/remotecommand:go_default_library",
         "//staging/src/k8s.io/client-go/tools/watch:go_default_library",
-        "//staging/src/k8s.io/client-go/util/retry:go_default_library",
         "//staging/src/k8s.io/component-base/cli/flag:go_default_library",
         "//test/e2e/framework/auth:go_default_library",
         "//test/e2e/framework/config:go_default_library",
@@ -153,6 +150,7 @@ filegroup(
         "//test/e2e/framework/providers/vsphere:all-srcs",
         "//test/e2e/framework/replicaset:all-srcs",
         "//test/e2e/framework/resource:all-srcs",
+        "//test/e2e/framework/service:all-srcs",
         "//test/e2e/framework/ssh:all-srcs",
         "//test/e2e/framework/statefulset:all-srcs",
         "//test/e2e/framework/testfiles:all-srcs",
