+++ test/soak/serve_hostnames/BUILD
@@ -26,6 +26,7 @@ go_library(
         "//staging/src/k8s.io/client-go/rest:go_default_library",
         "//staging/src/k8s.io/client-go/tools/clientcmd:go_default_library",
         "//test/e2e/framework:go_default_library",
+        "//test/e2e/framework/service:go_default_library",
         "//vendor/k8s.io/klog:go_default_library",
     ],
 )
